Add new API functions app_install() and app_uninstall().
The first function must be called by installer before any other call to app_*
functions for the app. It will create empty Smack config file for the app.
It will fail if the file was already there (conflicting app_ids).
Other functions operating on Smack rules for the application will fail if
this file is not present.
The latter function removes the file, supposed to be the last call to
libprivilege-control from uninstaller.
For temporary compatibility old permissive behaviour of app_add_permissions() is
retained.

[Issue#]       N/A
[Feature]      N/A
[Cause]        N/A
[Solution]     N/A
[Verification] Build, install, reboot target.

Change-Id: I371c38b88a0c14a17eade5526e8fa1276a482f85